proper hooking via vftable changed minhook wrap added source types refactors templated vector (duh)

+++ sdk/csgo/hooks/paint_traverse.hxx
@@ -1,11 +1,11 @@
 #pragma once
 
-using paint_traverse_t = std::int32_t( __fastcall* )( std::int32_t, std::int32_t, std::uint32_t, std::int8_t, std::int8_t );
+using paint_traverse_t = std::int32_t( __fastcall* )( std::ptrdiff_t, std::ptrdiff_t, std::uint32_t, std::int8_t, std::int8_t );
 static paint_traverse_t __paint_traverse;
 
 const std::int32_t __fastcall paint_traverse(
-   const std::int32_t ecx [[ maybe_unused ]],
-   const std::int32_t edx [[ maybe_unused ]],
+   const std::ptrdiff_t ecx [[ maybe_unused ]],
+   const std::ptrdiff_t edx [[ maybe_unused ]],
    const std::uint32_t panel,
    const std::int8_t repaint,
    const std::int8_t active
